move Hub stuff to own package -- make EventHandler public for tests
