+++ src/main/java/example/lsp/initialPlans/ExampleTwoEchelonGrid.java
@@ -23,6 +23,7 @@ package example.lsp.initialPlans;
 import lsp.*;
 import lsp.resourceImplementations.distributionCarrier.DistributionCarrierUtils;
 import lsp.resourceImplementations.mainRunCarrier.MainRunCarrierUtils;
+import lsp.resourceImplementations.transshipmentHub.TranshipmentHubUtils;
 import lsp.shipment.LSPShipment;
 import lsp.shipment.ShipmentUtils;
 import lsp.resourceImplementations.UsecaseUtils;
@@ -254,13 +255,13 @@ final class ExampleTwoEchelonGrid {
 					.build();
 
 			//The scheduler for the first reloading point is created --> this will be the depot in this use case
-			LSPResourceScheduler hubScheduler = UsecaseUtils.TranshipmentHubSchedulerBuilder.newInstance()
+			LSPResourceScheduler hubScheduler = TranshipmentHubUtils.TranshipmentHubSchedulerBuilder.newInstance()
 					.setCapacityNeedFixed(10) //Time needed, fixed (for Scheduler)
 					.setCapacityNeedLinear(1) //additional time needed per shipmentSize (for Scheduler)
 					.build();
 
 			//The scheduler is added to the Resource and the Resource is created
-			LSPResource hubResource = UsecaseUtils.TransshipmentHubBuilder.newInstance(Id.create("Hub", LSPResource.class), HUB_LINK_ID, scenario)
+			LSPResource hubResource = TranshipmentHubUtils.TransshipmentHubBuilder.newInstance(Id.create("Hub", LSPResource.class), HUB_LINK_ID, scenario)
 					.setTransshipmentHubScheduler(hubScheduler)
 					.build();
 			LSPUtils.setFixedCost(hubResource, HUBCOSTS_FIX); //Set fixed costs (per day) for the availability of the hub.
