+++ src/main/java/lsp/LSPPlanXmlWriter.java
@@ -23,7 +23,7 @@ package lsp;
 import lsp.shipment.LSPShipment;
 import lsp.shipment.ShipmentPlanElement;
 import lsp.shipment.ShipmentUtils;
-import lsp.resourceImplementations.TransshipmentHub;
+import lsp.resourceImplementations.transshipmentHub.TransshipmentHub;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.matsim.api.core.v01.Id;
